Added checkbox column to disable some paths

+++ Models/Functions.cs
@@ -2,24 +2,21 @@
 using System.ComponentModel;
 using System.IO;
 using System.Threading;
+using System.Windows;
 
 namespace FolderSyncer.Models
 {
-    class FolderToSync:INotifyPropertyChanged
+    class FolderToSync : INotifyPropertyChanged
     {
-        public FolderToSync(string p1, string p2,string time="")
+        public FolderToSync(string p1, string p2, string time = "")
         {
             PathFrom = p1;
             PathTo = p2;
             if (time == "")
-                LastUpdate = DateTime.Now.ToShortDateString()+" "+DateTime.Now.ToShortTimeString();
+                LastUpdate = DateTime.Now.ToShortDateString() + " " + DateTime.Now.ToShortTimeString();
             else
                 LastUpdate = time;
             BeginSyncing();
-        }
-        ~FolderToSync()
-        {
-            flag = false;
         }
         private Thread Syncer;
         private bool flag = true;
@@ -28,10 +25,20 @@ namespace FolderSyncer.Models
         {
             PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
         }
-
+        public bool Flag
+        {
+            get { return flag; }
+            set
+            {
+                flag = value;
+                OnPropertyChanged(nameof(Flag));
+                if (flag == true)
+                    BeginSyncing();
+            }
+        }
         public string PathFrom { get; set; }
         public string PathTo { get; set; }
-        private string _lp="";
+        private string _lp = "";
         public string LastUpdate
         {
             get { return _lp; }
@@ -68,21 +75,34 @@ namespace FolderSyncer.Models
             while (flag)
             {
                 string filename;
-                foreach (string file in Directory.GetFiles(PathFrom, "*", SearchOption.AllDirectories))
+                if (Directory.Exists(PathTo) && Directory.Exists(PathFrom))
                 {
-                    filename = file.Substring(PathFrom.Length, file.Length - PathFrom.Length);
-                    Sync(filename, CheckForChanges(filename));
-                }
-                foreach (string file in Directory.GetFiles(PathTo, "*", SearchOption.AllDirectories))
-                {
-                    filename = file.Substring(PathTo.Length, file.Length - PathTo.Length);
-                    if (!File.Exists(PathFrom + filename) && File.Exists(PathTo + filename) && !filename.Contains(".deleted"))
+                    foreach (string file in Directory.GetFiles(PathFrom, "*", SearchOption.AllDirectories))
                     {
-                        LastUpdate = DateTime.Now.ToShortDateString() + " " + DateTime.Now.ToShortTimeString();
-                        File.Move(PathTo + filename, PathTo + filename + ".deleted");
+                        filename = file.Substring(PathFrom.Length, file.Length - PathFrom.Length);
+                        Sync(filename, CheckForChanges(filename));
                     }
+                    foreach (string file in Directory.GetFiles(PathTo, "*", SearchOption.AllDirectories))
+                    {
+                        filename = file.Substring(PathTo.Length, file.Length - PathTo.Length);
+                        if (!File.Exists(PathFrom + filename) && File.Exists(PathTo + filename) && !filename.Contains(".deleted"))
+                        {
+                            LastUpdate = DateTime.Now.ToShortDateString() + " " + DateTime.Now.ToShortTimeString();
+                            File.Move(PathTo + filename, PathTo + filename + ".deleted");
+                        }
+                    }
+                    Thread.Sleep(1000);
                 }
-                Thread.Sleep(1000);
+                else if(!Directory.Exists(PathTo))
+                {
+                    MessageBox.Show("Directory "+PathTo+" wasn't found","Error");
+                    Flag = false;
+                }
+                else if (!Directory.Exists(PathFrom))
+                {
+                    MessageBox.Show("Directory " + PathFrom + " wasn't found", "Error");
+                    Flag = false;
+                }
             }
 
         }
